extract smoothScroll to scrollManager to reduce vimState file size
