+++ lib/demo-mode-support.js
@@ -1,13 +1,14 @@
 const globalState = require("./global-state")
 const VimState = require("./vim-state")
 const Base = require("./base")
+const {CompositeDisposable} = require("atom")
 
 class DemoModeSupport {
   init({onWillAddItem, onDidStart, onDidStop, onWillFadeoutHover, onDidRemoveHover}) {
     const destroyAllDemoModeFlasheMarkers = this.destroyAllDemoModeFlasheMarkers.bind(this)
 
     // Returns disposables
-    return [
+    return new CompositeDisposable(
       onDidStart(() => globalState.set("demoModeIsActive", true)),
       onDidStop(() => globalState.set("demoModeIsActive", false)),
       onDidRemoveHover(destroyAllDemoModeFlasheMarkers),
@@ -21,8 +22,8 @@ class DemoModeSupport {
         element.classList.add("kind", "pull-right")
         element.textContent = this.getKindForCommand(event.binding.command)
         item.appendChild(element)
-      }),
-    ]
+      })
+    )
   }
 
   destroyAllDemoModeFlasheMarkers() {
