+++ lib/main.js
@@ -187,8 +187,7 @@ module.exports = {
   },
 
   consumeDemoMode(service) {
-    const demoModeSupport = require("./demo-mode-support")
-    this.subscriptions.add(...demoModeSupport.init(service))
+    this.subscriptions.add(this.demoModeSupport.init(service))
   },
 
   // Computed props
@@ -196,11 +195,12 @@ module.exports = {
   get statusBarManager() {
     return this.__statusBarManager || (this.__statusBarManager = require("./status-bar-manager"))
   },
-
+  get demoModeSupport() {
+    return this.__demoModeSupport || (this.__demoModeSupport = require("./demo-mode-support"))
+  },
   get paneUtils() {
     return this.__paneUtils || (this.__paneUtils = require("./pane-utils"))
   },
-
   get globalState() {
     return this.__globalState || (this.__globalState = require("./global-state"))
   },
@@ -210,7 +210,6 @@ module.exports = {
   getGlobalState() {
     return this.globalState
   },
-
   getEditorState(editor) {
     return VimState.get(editor)
   },
