+++ lib/utils.js
@@ -1123,24 +1123,6 @@ function getRowRangeForCommentAtBufferRow(editor, row) {
   return [startRow, endRow]
 }
 
-function animateInDuration(msec, fn) {
-  let startTime, nextRequestID
-
-  function step(timestamp) {
-    if (!startTime) startTime = timestamp
-    const progress = Math.min((timestamp - startTime) / msec, 1)
-    if (progress < 1) {
-      nextRequestID = window.requestAnimationFrame(step)
-    }
-    fn(progress)
-  }
-  nextRequestID = window.requestAnimationFrame(step)
-
-  return () => {
-    window.cancelAnimationFrame(nextRequestID)
-  }
-}
-
 module.exports = {
   assertWithException,
   getLast,
@@ -1238,5 +1220,4 @@ module.exports = {
   normalizeIndent,
   atomVersionSatisfies,
   getRowRangeForCommentAtBufferRow,
-  animateInDuration,
 }
