+++ lib/vim-state.js
@@ -50,6 +50,7 @@ module.exports = class VimState {
   get operationStack() { return this.__operationStack || (this.__operationStack = this.load("./operation-stack")) } // prettier-ignore
   get cursorStyleManager() { return this.__cursorStyleManager || (this.__cursorStyleManager = this.load("./cursor-style-manager")) } // prettier-ignore
   get sequentialPasteManager() { return this.__sequentialPasteManager || (this.__sequentialPasteManager = this.load("./sequential-paste-manager")) } // prettier-ignore
+  get scrollManager() { return this.__scrollManager || (this.__scrollManager = this.load("./scroll-manager")) } // prettier-ignore
   get swrap() { return this.__swrap || (this.__swrap = this.load("./selection-wrapper", false)) } // prettier-ignore
   get utils() { return this.__utils || (this.__utils = this.load("./utils", false)) } // prettier-ignore
   get globalState() { return this.__globalState || (this.__globalState = this.load("./global-state", false)) } // prettier-ignore
@@ -66,7 +67,6 @@ module.exports = class VimState {
     this.submode = null
 
     this.previousSelection = {}
-    this.scrollRequest = null
     this.ignoreSelectionChange = false
 
     this.subscriptions = new CompositeDisposable(
@@ -406,53 +406,8 @@ module.exports = class VimState {
     if (this.__persistentSelection) this.persistentSelection.clearMarkers()
   }
 
-  // It's possible that multiple scroll request comes in sequentially.
-  //
-  // When you pass `amountOfPixels`, it calculate final scrollTop based on
-  // prevoiusly requested scrollTop. In other word. animation is concatnated to
-  // scroll smoothly in sequential request.
-  //
-  // When you pass `scrollTop`, it's absolute value, just use passed scrollTop.
-  // So passed scrollTop should not be based on **relative** scrollTop, because it's might not finalized.
-  // - OK: {scrollTop: cursor.pixelPositionForScreenPosition(point) + pixels}
-  // - NG: {scrollTop: editorElement.getScrollTop() + pixels}
-  // - Hacky but should be OK: {scrollTop: vimState.scrollRequest.scrollTop + pixels}
-  requestScroll({amountOfPixels, scrollTop, duration, onFinish}) {
-    const currentScrollTop = this.editorElement.getScrollTop()
-    let baseScrollTop = currentScrollTop
-
-    if (this.scrollRequest) {
-      this.scrollRequest.cancel()
-      baseScrollTop = this.scrollRequest.scrollTop
-      this.scrollRequest = null
-    }
-
-    if (amountOfPixels) {
-      scrollTop = baseScrollTop + amountOfPixels
-    }
-
-    if (!duration) {
-      this.editorElement.setScrollTop(scrollTop)
-      if (onFinish) onFinish()
-      return
-    }
-
-    const deltaInPixels = scrollTop - currentScrollTop
-    const cancel = this.utils.animateInDuration(duration, progress => {
-      if (this.editorElement.component) {
-        // [NOTE]
-        // intentionally use `element.component.setScrollTop` instead of `element.setScrollTop`.
-        // Since element.setScrollTop will throw exception when element.component no longer exists.
-        this.editorElement.component.setScrollTop(currentScrollTop + deltaInPixels * progress)
-        this.editorElement.component.updateSync()
-      }
-      if (progress === 1) {
-        this.scrollRequest = null
-        if (onFinish) onFinish()
-      }
-    })
-
-    this.scrollRequest = {scrollTop, cancel}
+  requestScroll(...args) {
+    this.scrollManager.requestScroll(...args)
   }
 
   // Mode Managerment
